+++ src/Ext/Techno/Hooks.Firing.cpp
@@ -1268,3 +1268,26 @@ DEFINE_HOOK(0x4D5A34, FootClass_ApproachTarget_StopWhenInRange, 0x6)
 
 	return 0;
 }
+
+DEFINE_HOOK(0x4D57EA, FootClass_ApproachTarget_PursuitTarget, 0x9)
+{
+	enum { Return = 0x4D5A34 };
+
+	GET(FootClass*, pThis, EBX);
+	GET_STACK(const bool, closeEnough, STACK_OFFSET(0x158, -0x146));
+
+	const auto pTypeExt = TechnoExt::Fetch(pThis)->TypeExtData;
+
+	if (closeEnough && pTypeExt->ApproachTarget_StopWhenInRange.Get(RulesExt::Global()->ApproachTarget_StopWhenInRange))
+		return 0;
+
+	if (pTypeExt->ApproachTarget_PursuitTarget)
+	{
+		pThis->SetDestination(pThis->Target, true);
+		R->EDI(0);
+		R->Stack(STACK_OFFSET(0x158, -0x130), 0);
+		return Return;
+	}
+
+	return 0;
+}